refactor: move report code to report pages dir
