+++ src/pages/report/report/reportParts/CoverPage.scss
