+++ src/pages/report/report/reportParts/ErrorPage.scss
