+++ src/pages/report/report/reportParts/Group.scss
