+++ src/pages/report/report/reportParts/Hazard.scss
