+++ src/pages/report/report/reportParts/HazardMap.scss
