+++ src/pages/report/report/reportParts/HazardMap.tsx
