+++ src/pages/report/report/reportParts/HazardUnit.scss
