+++ src/pages/report/report/reportParts/LidarFeature.scss
