+++ src/pages/report/report/reportParts/Loader.scss
