+++ src/pages/report/report/reportParts/OtherDataPage.scss
