+++ src/pages/report/report/reportParts/ProgressBar.scss
