+++ src/pages/report/report/reportParts/ProgressBar.tsx
