+++ src/pages/report/report/variables.scss
